+++ src/conversion/pipelineJsonToString.ts
@@ -164,5 +164,8 @@ function promptTemplateParameterJsonToString(promptTemplateParameterJson: Prompt
 }
 
 /**
+ * TODO: !!!! Auto detect missing features in pipelineJsonToString
+ * TODO: !!!! Knowledge and samples
  * TODO: Escape all
+ * TODO: [😵] Use markdownStructureToMarkdown
  */